[test] refactor: split apply_veomni_hf_unpatch into a leaner module
`tests/models/utils.py` imports `veomni.data.dummy_dataset` at module top,
which transitively pulls `av`/`torchcodec` through the multimodal data
package init. That makes any test that just needs `apply_veomni_hf_unpatch`
unable to be collected without the optional video extra installed.

Move the pristine HF class snapshot + `apply_veomni_hf_unpatch` into a
sibling `tests/models/hf_unpatch.py` whose only imports are `transformers`
modeling modules. `utils.py` re-exports `apply_veomni_hf_unpatch` so existing
call sites stay unchanged. The new logits-equal smoke test now collects in
environments without `av` installed.
